pyproject.toml
[project]
name = "uv-package-template"
version = "0.1.0"
description = "Template for UV package"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.0.1",
]

[project.scripts]
main = "uv_package_template.cli:main"
alt = "uv_package_template.cli:alt"

[project.optional-dependencies]
dev = [
  "ruff>=0.12.10",
  "pytest>=8.3.2",
  "mypy>=1.10.0",
  "poethepoet>=0.26.1",
]

[tool.poe.tasks]
lint = "ruff check ."
fmt = "ruff format"
typecheck = "mypy --sqlite-cache --pretty src tests"
test = "pytest -q"
check = { sequence = ["lint", "typecheck", "test"] }

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uv_package_template"]

[tool.ruff]
target-version = "py313"
line-length = 100
fix = true

[tool.ruff.lint]
# Core codes: pyflakes/pycodestyle/imports/bugbear/pyupgrade/simplify + Ruff rules.
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
disallow_untyped_defs = true
check_untyped_defs = true
pretty = true
